perf: remove unneeded type change
Row names are not being saved anyways so there is no need to convert to
data table to keep this information in the resultant matrix. The states
are already sorted by this point, so there is no need. States shouldn't
ever be missed out here as it would imply that the state is never
actually assigned to a bin by ChromHMM (which almost certainly means the
model is too complex).