feat: Add spatial similarity (#12)
* feat: add script to create a blank bed file with constant size

* feat: add creation of blank files to main script

* feat: add chromosome sizes file to config

* feat: add bedtools to convert segment files to equidistant bins

I wish ChromHMM did this for me, but hey ho.

* refactor: split current spatial comparison pipeline into further
functions

* chore: add explanation behind converting of state assignments files

* refactor: remove unused parameter

* feat: change shell script to accomodate new approach

* refactor: rename function to be consistent

* chore: add comment explaining sorting

* fix: use correct function arguments

* refactor: change prefix of file path

* feat: add skeleton for adding margins

* feat: add the margin adding function

* feat: add function to get chromosome sizes

* feat: add function to correct bounds if they are nonsensicle

* fix: ensure correcting only occurs for specific chromosome

* perf: start bp being wrong is not chromosome specific

* feat: add file saving function

* fix: add output file path as argument

* fix: & doesn't work here, nest `ifelse` instead

* fix: force R to use uint

* feat: use bedtools to get desired form for R script

Idea is to have a file that is just a list of each state pair that
overlap with one another. The fold enrichment will be incredibly simple
and quick now to calculate. Meaning most of the R script can focus on
putting this into a nice form

* feat: add bin size (necessary command line argument)

* feat: add outline for spatial similarity script

* perf: reimagine stats table to reduce repeated calculations

* feat: read in combined assignments file

* feat: add function to add states column

* feat: add function to calculate state coverage

* fix: get the stats table generation functions to play nice

* feat: include amount of overlap in combined state assignments

* revert: remove heatmap from skeleton

I'd rather just make a supplementary script at some point that users can
use to convert any matrix into a heatmap themselves.

* feat: add function to combine stats tables

* feat: add function to calculate genome size

Genome size is just the bin size times the total number of lines in the
state assignment file

* fix: column names are not required as already set when merging

* feat: add function to add overlap to stats table

* perf: use grouping and left join to avoid double loop

* feat: add functions to add fold_enrichment

* feat: add function to convert data table into matrix form

* feat: add function for saving matrix

* perf: remove unneeded type change

Row names are not being saved anyways so there is no need to convert to
data table to keep this information in the resultant matrix. The states
are already sorted by this point, so there is no need. States shouldn't
ever be missed out here as it would imply that the state is never
actually assigned to a bin by ChromHMM (which almost certainly means the
model is too complex).

* fix: typo

* fix: bp_coverage columns were getting lost

Just left join again to bring back to the party

* fix: avoid coercion to data table

* fix: typo

* fix: pass only the one column

Passing both columns results in sum() counting instances of the state in
both models (not required behaviour)