refactor: simplify bash array validation
There's too much shit that can happen inside of the bash array that I
cannot account for with a simple bash array. On the subject of
variables, if I did allow for them and check these are valid. How would
I know that the actual values are correct? I'm just going to ignore
this.