feat: Add config checker (#15)
* feat: add skeleton of config file checker

* feat: add eol validation

Config file must be in linux eol format as this is what the rest of the
pipeline implicitly depends on. A different eol format would result in
the `source` commands to fail.

* fix: robustly call python script

Cannot depend on the config file as it hasn't been checked or sourced by
this point.

* feat: ensure that script is run via sbatch

If the script is ran just with bash, then the changing directory logic
will break.

* fix: remove uncommented lines

* feat: add function to extract all config variables

* feat: ensure that variables have values

* fix: PYTHON_DIRECTORY unused

* feat: add function to show all variables exist

* feat: add validation of variable types

* feat: implement error messaging into pipeline

* feat: add check to ensure all paths are correct

This doesn't account for the off chance that users will be lazy and use
variables to create their paths.

* refactor: simplify bash array validation

There's too much shit that can happen inside of the bash array that I
cannot account for with a simple bash array. On the subject of
variables, if I did allow for them and check these are valid. How would
I know that the actual values are correct? I'm just going to ignore
this.

* fix: add proper spacing to print statements

* feat: add validation for the size of the weights and margins

* fix: ensure PROCESSING DIRECTORY is defined